backend/requirements.txt
Django==5.2.1
mysqlclient
djangorestframework
django-cors-headers
